requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.2
Werkzeug==2.3.4
openai==0.27.8
python-dotenv==1.0.0
groq